fix circular dependency hell ;D [3/3]; change database URI environment variable from the config.py file fix script to run the backend server from the project root todo: test database endpoints
