+++ backend/config.py
@@ -8,7 +8,7 @@ class Config:
     # this is used, but the best practice is to set environment
     # variables or secrets to handle these infos.
     SECRET_KEY = 'gV7S6iGV560ZikWik170'
-    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URI') or \
+    SQLALCHEMY_DATABASE_URI = os.environ.get('DB_URL') or \
         'postgresql://' + os.path.join(basedir, 'app.db')
     SQLALCHEMY_TRACK_MODIFICATIONS = False
     STATIC_FOLDER = f"{os.getenv('APP_FOLDER')}/project/static"
